Add documentation in .Rd files
At one point we had the notion of submitting this package to CRAN. The
CRAN reviewer requested adding at least one .Rd file to document the
package.

Even though we no-longer plan to submit to CRAN, this seems like a good
suggestion.

Also added quick documentation for onAttach and onLoad.